+++ src/examples/ApostaExample.ts
@@ -13,14 +13,9 @@ export const criarApostas = async (connection: Connection) => {
     const partidaRepo = connection.getCustomRepository(PartidaRepository);
     const apostaRepo = connection.getCustomRepository(ApostaRepository);
     const usuarioRepo = connection.getCustomRepository(UsuarioRepository);
-    const campeonatoRepo = connection.getCustomRepository(CampeonatoRepository);
 
     const usuarioService = new UsuarioService(usuarioRepo);
     const apostaService = new ApostaService(apostaRepo, usuarioRepo, partidaRepo, rodadaRepo);
-
-    const usuarioBD = await usuarioRepo.findByEmail("erik@email.com");
-    const campeonato = await campeonatoRepo.findBySlug("brasileirão-2021");
-    usuarioService.adicionarCampeonato(usuarioBD.id, campeonato);
 
     const usuarioLogado = await usuarioService.autenticar({
         email: "erik@email.com",
@@ -46,5 +41,40 @@ export const criarApostas = async (connection: Connection) => {
     const usuario = await usuarioRepo.findByEmail(usuarioLogado.email);
 
     await apostaService.gerarApostas(usuario.id, 1, palpites);
-
 };
+
+export const adicionarUsuarioAoCampeonato = async (connection: Connection) => {
+    const usuarioRepo = connection.getCustomRepository(UsuarioRepository);
+    const campeonatoRepo = connection.getCustomRepository(CampeonatoRepository);
+
+    const usuarioService = new UsuarioService(usuarioRepo);
+
+    const usuarioBD = await usuarioRepo.findByEmail("erik@email.com");
+    const campeonato = await campeonatoRepo.findBySlug("brasileirão-2021");
+    usuarioService.adicionarCampeonato(usuarioBD.id, campeonato);
+}
+
+export const atualizarPontuacao = async (connection: Connection) => {
+    const rodadaRepo = connection.getCustomRepository(RodadaRepository);
+    const partidaRepo = connection.getCustomRepository(PartidaRepository);
+    const apostaRepo = connection.getCustomRepository(ApostaRepository);
+    const usuarioRepo = connection.getCustomRepository(UsuarioRepository);
+
+    const apostaService = new ApostaService(apostaRepo, usuarioRepo, partidaRepo, rodadaRepo);
+
+    await apostaService.atualizarPontuacao();
+};
+
+export const atualizarClassificacao = async (connection: Connection) => {
+    const rodadaRepo = connection.getCustomRepository(RodadaRepository);
+    const partidaRepo = connection.getCustomRepository(PartidaRepository);
+    const apostaRepo = connection.getCustomRepository(ApostaRepository);
+    const usuarioRepo = connection.getCustomRepository(UsuarioRepository);
+
+    const apostaService = new ApostaService(apostaRepo, usuarioRepo, partidaRepo, rodadaRepo);
+
+    const classificacao = await apostaService.gerarClassificacao();
+    classificacao.forEach((posicao, i) => {
+        console.log(`${i+1}° ${posicao.usuario.nome} - ${posicao.pontuacao} ponto(s)`);
+    });
+}