+++ src/repositories/ApostaRepository.ts
@@ -4,6 +4,11 @@ import { IApostaRepository } from "./IApostaRepository";
 
 @EntityRepository(Aposta)
 export class ApostaRepository extends Repository<Aposta> implements IApostaRepository {
+    findAll(): Promise<Aposta[]> {
+        return this.find({
+            relations: ['partida', 'usuario']
+        })
+    }
     findbyUsuarioAndPartida(usuarioId: number, partidaId: number): Promise<Aposta> {
         return this.findOne({
             where: {