+++ src/service/__tests__/ApostaService.test.ts
@@ -13,16 +13,20 @@ import { Aposta } from '../../models/ApostaEntity';
 
 describe("ApostaService", () => {
     let usuarioId: number;
+    let usuarioId2: number;
     let partidaId: number;
     let campeonatoId: number;
     let numeroRodada: number;
     let rodadaId: number;
     let palpite: PalpiteDto;
     let usuario: Usuario;
+    let usuario2: Usuario;
     let rodada: Rodada;
     let campeonato: Campeonato;
     let partida: Partida;
     let aposta: Aposta;
+    let apostaPontuada: Aposta;
+    let apostaPontuada2: Aposta;
 
     let apostaRepo: ApostaRepository;
     let rodadaRepo: RodadaRepository;
@@ -38,6 +42,7 @@ describe("ApostaService", () => {
         apostaService = new ApostaService(apostaRepo, usuarioRepo, partidaRepo, rodadaRepo);
 
         usuarioId = faker.datatype.number();
+        usuarioId2 = faker.datatype.number();
         campeonatoId = faker.datatype.number();
         partidaId = faker.datatype.number();
         rodadaId = faker.datatype.number();
@@ -59,6 +64,13 @@ describe("ApostaService", () => {
         usuario.id = usuarioId;
         usuario.campeonatos = [];
 
+        usuario2 = new Usuario();
+        usuario2.nome = faker.lorem.word();
+        usuario2.hashSenha = faker.internet.password();
+        usuario2.email = faker.internet.email();
+        usuario2.id = faker.datatype.number();
+        usuario2.campeonatos = [];
+
         rodada = new Rodada();
         rodada.campeonato = campeonato;
         rodada.nome = faker.lorem.word();
@@ -69,10 +81,12 @@ describe("ApostaService", () => {
 
         partida = new Partida();
         partida.id = partidaId;
+        partida.placarMandante = faker.datatype.number();
+        partida.placarVisitante = faker.datatype.number();
         
         palpite = {
-            golsMandante: faker.datatype.number(),
-            golsVisitante: faker.datatype.number(),
+            golsMandante: partida.placarMandante,
+            golsVisitante: partida.placarVisitante,
             partidaID: partidaId,
         }
 
@@ -81,10 +95,20 @@ describe("ApostaService", () => {
         aposta.placarMandante = palpite.golsMandante;
         aposta.placarVisitante = palpite.golsVisitante;
         aposta.usuario = usuario;
+
+        apostaPontuada = aposta;
+        apostaPontuada.pontos = faker.datatype.number();
+
+        apostaPontuada2 = new Aposta();
+        apostaPontuada2.usuario = usuario2;
+        apostaPontuada2.partida = partida;
+        apostaPontuada2.placarMandante = palpite.golsMandante;
+        apostaPontuada2.placarVisitante = palpite.golsVisitante;
+        apostaPontuada2.pontos = faker.datatype.number();
     });
 
     beforeEach(jest.clearAllMocks);
-    describe("criarApostas", () => {
+    describe("gerarApostas", () => {
         it("Deve retornar um erro caso não consiga buscar o usuário no banco de dados", async () => {
             jest.spyOn(usuarioRepo, "findById").mockRejectedValue(new Error("Não foi possivel buscar dados do usuario no banco"));
 
@@ -158,6 +182,66 @@ describe("ApostaService", () => {
             jest.spyOn(apostaRepo, "findbyUsuarioAndPartida").mockResolvedValue(aposta);
 
             await expect(apostaService.gerarApostas(usuarioId, numeroRodada, [palpite])).resolves.not.toBeDefined();
+        });
+    });
+    describe("atualizarPontuacao", () => {
+        it("Deve retornar um erro caso não consiga buscar as apostas do banco", async () => {
+            jest.spyOn(apostaRepo, "findAll").mockRejectedValue(new Error("Não foi possivel buscar dado das apostas no banco"));
+
+            await expect(apostaService.atualizarPontuacao()).rejects.toThrow();
+        });
+        it("Deve retornar um erro caso não consiga buscar as apostas do banco", async () => {
+            jest.spyOn(apostaRepo, "findAll").mockResolvedValue([aposta]);
+            jest.spyOn(apostaRepo, "save").mockRejectedValue(new Error("Não foi possivel salvar as apostas no banco"))
+
+            await expect(apostaService.atualizarPontuacao()).rejects.toThrow();
+        });
+        it("Não deve atualizar apostas de jogos não finalizados", async () => {
+            partida.status = faker.lorem.word();
+            jest.spyOn(apostaRepo, "findAll").mockResolvedValue([aposta]);
+            jest.spyOn(apostaRepo, "save").mockResolvedValue(aposta);
+
+            await expect(apostaService.atualizarPontuacao()).resolves.not.toBeDefined();
+            expect(apostaRepo.save).toBeCalledWith(undefined);
+        });
+        it("Deve atualizar corretamente a pontuação de apostas no banco em jogos finalizados", async () => {
+            partida.status = "finalizado";
+            jest.spyOn(apostaRepo, "findAll").mockResolvedValue([aposta]);
+            jest.spyOn(apostaRepo, "save").mockResolvedValue(aposta);
+            
+            await expect(apostaService.atualizarPontuacao()).resolves.not.toBeDefined();
+            expect(apostaRepo.save).toBeCalledWith(expect.objectContaining({pontos: expect.any(Number)}));
+        });
+        it("Deve atualizar corretamente a pontuação de apostas no banco em jogos finalizados, cenário para erro em todos os casos", async () => {
+            partida.placarMandante = faker.datatype.number();
+            partida.placarVisitante = faker.datatype.number(1);
+            partida.status = "finalizado";
+            jest.spyOn(apostaRepo, "findAll").mockResolvedValue([aposta]);
+            jest.spyOn(apostaRepo, "save").mockResolvedValue(aposta);
+            
+            await expect(apostaService.atualizarPontuacao()).resolves.not.toBeDefined();
+            expect(apostaRepo.save).toBeCalledWith(expect.objectContaining({pontos: expect.any(Number)}));
+        });
+    });
+    describe("gerarClassificacao", () => {
+        it("Deve retornar um erro caso não consiga buscar as apostas do banco", async () => {
+            jest.spyOn(apostaRepo, "findAll").mockRejectedValue(new Error("Não foi possivel buscar dado das apostas no banco"));
+
+            await expect(apostaService.gerarClassificacao()).rejects.toThrow();
+        });
+        it("Deve retornar um erro caso não consiga buscar os usuarios no banco", async () => {
+            jest.spyOn(apostaRepo, "findAll").mockResolvedValue([apostaPontuada]);
+            jest.spyOn(usuarioRepo, "findById").mockRejectedValue(new Error("Não foi possivel buscar dado das apostas no banco"));
+
+            await expect(apostaService.gerarClassificacao()).rejects.toThrow();
+        });
+        it("Deve retornar a classificação", async () => {
+            jest.spyOn(apostaRepo, "findAll").mockResolvedValue([apostaPontuada, apostaPontuada2]);
+            jest.spyOn(usuarioRepo, "findById").mockResolvedValueOnce(usuario);
+            jest.spyOn(usuarioRepo, "findById").mockResolvedValueOnce(usuario2);
+
+            const resposta = await expect(apostaService.gerarClassificacao()).resolves.toBeDefined();
+            expect(resposta).not.toBeNull();
         });
     });
 });